Refactor withTableDropdown hoc
