+++ packages/react/src/index.ts
@@ -7,10 +7,10 @@ export * from './ArqueroTableHeader'
 export * from './ArqueroTable'
 export * from './verbs'
 export * from './controls/dropdowns/TableDropdown'
-export * from './WithTableDropdown'
 export * from './VerbDescription'
 export * from './common'
 export * from './selectStepComponent'
 export * from './selectStepDescription'
+export * from './withTableDropdown'
 export * from './types'
 export * from './files'
